+++ apps/fabric-tester/android/app/src/main/java/com/community/fabrictester/MainActivity.java
@@ -40,10 +40,7 @@ public class MainActivity extends ReactActivity {
         this,
         getMainComponentName(),
         // If you opted-in for the New Architecture, we enable the Fabric Renderer.
-        DefaultNewArchitectureEntryPoint.getFabricEnabled(), // fabricEnabled
-        // If you opted-in for the New Architecture, we enable Concurrent React (i.e. React 18).
-        DefaultNewArchitectureEntryPoint.getConcurrentReactEnabled() // concurrentRootEnabled
-        ));
+        DefaultNewArchitectureEntryPoint.getFabricEnabled()));
   }
 
   /**